Move onCreateUser to shared
