+++ backend/scripts/compare-bet-endpoints.ts
@@ -7,13 +7,12 @@ import { getRandomTestBet } from 'shared/test/bets'
 import { ValidatedAPIParams } from 'common/api/schema'
 import { ContractMetric } from 'common/contract-metric'
 import { convertContract } from 'common/supabase/contracts'
-import { DEV_CONFIG } from 'common/envs/dev'
 import { floatingEqual } from 'common/util/math'
 
-const API_URL = `https://${DEV_CONFIG.apiEndpoint}/v0`
-// const API_URL = `http://localhost:8088/v0`
-const BETS_PER_USER_PER_MARKET = 1
-const NUM_USERS = 160
+// const API_URL = `https://${DEV_CONFIG.apiEndpoint}/v0`
+const API_URL = `http://localhost:8088/v0`
+const BETS_PER_USER_PER_MARKET = 10
+const NUM_USERS = 10
 const ENABLE_LIMIT_ORDERS = true
 
 async function createTestMarkets(pg: SupabaseDirectClient, apiKey: string) {
@@ -124,15 +123,19 @@ async function comparePositions(
 
   for (const [userId, pos1] of posMap1) {
     const pos2 = posMap2.get(userId)!
-    const unequalPos = Object.entries(pos1.totalShares).find(
-      ([key, value]) => !floatingEqual(pos2.totalShares[key], value)
+    const unequalShares = Object.keys(pos1.totalShares).filter(
+      (key) => !floatingEqual(pos1.totalShares[key], pos2.totalShares[key])
     )
-    if (!pos2 || pos1.payout !== pos2.payout || unequalPos) {
+    if (!pos2 || pos1.payout !== pos2.payout || unequalShares.length > 0) {
       log(
         `Mismatch for user ${userId} in markets ${market1.id} and ${market2.id}`
       )
-      if (unequalPos) {
-        log(`Unequal position: ${unequalPos}`)
+      if (unequalShares.length > 0) {
+        unequalShares.forEach((key) => {
+          log(`Unequal position for ${key}:`)
+          log(`  Market 1: ${pos1.totalShares[key]}`)
+          log(`  Market 2: ${pos2.totalShares[key]}`)
+        })
       }
       log(`Market 1: ${JSON.stringify(pos1)}`)
       log(`Market 2: ${JSON.stringify(pos2)}`)
@@ -182,45 +185,57 @@ if (require.main === module) {
           answerId: multiChoiceMarkets[1].answers[answerIndex].id,
         }
 
-        const bets = [
-          { type: 'bet', bet: binaryBet },
-          { type: 'bet-batched', bet: binaryBet2 },
-          { type: 'bet', bet: multiChoiceBet },
-          { type: 'bet-batched', bet: multiChoiceBet2 },
+        const bothBets = [
+          [
+            { type: 'bet', bet: binaryBet },
+            { type: 'bet-batched', bet: binaryBet2 },
+          ],
+          [
+            { type: 'bet', bet: multiChoiceBet },
+            { type: 'bet-batched', bet: multiChoiceBet2 },
+          ],
         ]
-
-        for (const { type, bet } of bets) {
-          let betPlaced = false
-          let attempts = 0
-          while (!betPlaced) {
-            attempts++
-            totalAttempts++
-            try {
-              await placeBet(type, bet, user.apiKey)
-              betPlaced = true
-              successfulBets++
-            } catch (error: any) {
-              if ('message' in error) {
-                if (error.message === 'Betting allowed only between 1-99%') {
-                  break
+        await Promise.all(
+          bothBets.map(async (bets) => {
+            let skipBetPair = false
+            for (const { type, bet } of bets) {
+              if (skipBetPair) break
+              let betPlaced = false
+              let attempts = 0
+              while (!betPlaced) {
+                attempts++
+                totalAttempts++
+                try {
+                  await placeBet(type, bet, user.apiKey)
+                  betPlaced = true
+                  successfulBets++
+                } catch (error: any) {
+                  if ('message' in error) {
+                    if (
+                      error.message === 'Betting allowed only between 1-99%'
+                    ) {
+                      skipBetPair = true
+                      break
+                    }
+                    log(`Bet failed (attempt ${attempts}): ${error.message}`)
+                  } else log(`Bet failed (attempt ${attempts}): ${error}`)
+                  if (attempts % 5 === 0) {
+                    log(`Continuing to retry for bet: ${JSON.stringify(bet)}`)
+                  }
                 }
-                log(`Bet failed (attempt ${attempts}): ${error.message}`)
-              } else log(`Bet failed (attempt ${attempts}): ${error}`)
-              if (attempts % 5 === 0) {
-                log(`Continuing to retry for bet: ${JSON.stringify(bet)}`)
               }
             }
-          }
-        }
-        if (successfulBets % 8 === 0) {
-          const currentTime = Date.now()
-          const elapsedTime = (currentTime - startTime) / 1000
-          log(
-            `In ${elapsedTime.toFixed(2)} seconds:
+            if (successfulBets % 8 === 0) {
+              const currentTime = Date.now()
+              const elapsedTime = (currentTime - startTime) / 1000
+              log(
+                `In ${elapsedTime.toFixed(2)} seconds:
                 ${successfulBets} successful bets,
                 ${totalAttempts} total attempts`
-          )
-        }
+              )
+            }
+          })
+        )
       }
     }
