+++ common/src/api/schema.ts
@@ -192,9 +192,12 @@ export const API = (_apiTypeCheck = {
         replyToCommentId: z.string().optional(),
         limitProb: z.number().gte(0.01).lte(0.99).optional(),
         expiresAt: z.number().optional(),
+        // Used for binary and new multiple choice contracts (cpmm-multi-1).
         outcome: z.enum(['YES', 'NO']).default('YES'),
+        //Multi
         answerId: z.string().optional(),
         dryRun: z.boolean().optional(),
+        deps: z.array(z.string()).optional(),
       })
       .strict(),
   },
